ml/requirements.txt
fastapi
uvicorn
python-multipart
numpy==1.23.5
pillow==10.3.0
keras==2.12.0
tensorrt
tensorflow==2.12.0
tf_keras
